Fix sending `self error: msg`
The previous implementation would replace the `at:` with an `error:`. So, on the first error, instead of just reporting the error, from then on, it would always report the error even when it should just do a normal `at:`.

I am using the same approach here as SAbstractObject.send* methods.

It is also important to avoid concatinating strings on the fast path. The partial evaluation can’t handle that.

Signed-off-by: Stefan Marr <git@stefan-marr.de>